[build] move spinel into src/utils
Move spinel from src/ncp into src/utils and remove dependency on
OpenThread's core configurations.
