+++ src/ncp/Makefile.am
@@ -62,6 +62,7 @@ libopenthread_ncp_mtd_a_CPPFLAGS                  = \
     -DOPENTHREAD_RADIO=0                            \
     -DOPENTHREAD_FTD=0                              \
     -DOPENTHREAD_MTD=1                              \
+    -DOPENTHREAD_MESSAGE_IN_SPINEL_BUFFER=1         \
     $(COMMON_CPPFLAGS)                              \
     $(NULL)
 
@@ -69,6 +70,7 @@ libopenthread_ncp_ftd_a_CPPFLAGS                  = \
     -DOPENTHREAD_RADIO=0                            \
     -DOPENTHREAD_FTD=1                              \
     -DOPENTHREAD_MTD=0                              \
+    -DOPENTHREAD_MESSAGE_IN_SPINEL_BUFFER=1         \
     $(COMMON_CPPFLAGS)                              \
     $(NULL)
 
@@ -76,8 +78,18 @@ libopenthread_rcp_a_CPPFLAGS                      = \
     -DOPENTHREAD_RADIO=1                            \
     -DOPENTHREAD_FTD=0                              \
     -DOPENTHREAD_MTD=0                              \
+    -DOPENTHREAD_MESSAGE_IN_SPINEL_BUFFER=0         \
     $(COMMON_CPPFLAGS)                              \
     $(NULL)
+
+libopenthread_ncp_mtd_a_LIBADD                    = \
+    $(addprefix $(top_builddir)/src/utils/,$(filter %.o,$(shell $(AR) t $(top_builddir)/src/utils/libopenthread-spinel-ncp.a)))
+
+libopenthread_ncp_ftd_a_LIBADD                    = \
+    $(addprefix $(top_builddir)/src/utils/,$(filter %.o,$(shell $(AR) t $(top_builddir)/src/utils/libopenthread-spinel-ncp.a)))
+
+libopenthread_rcp_a_LIBADD                        = \
+    $(addprefix $(top_builddir)/src/utils/,$(filter %.o,$(shell $(AR) t $(top_builddir)/src/utils/libopenthread-spinel-rcp.a)))
 
 COMMON_SOURCES                                    = \
     changed_props_set.cpp                           \
@@ -90,20 +102,11 @@ COMMON_SOURCES                                    = \
     ncp_base_mtd.cpp                                \
     ncp_base_radio.cpp                              \
     ncp_base_dispatcher.cpp                         \
-    ncp_buffer.cpp                                  \
-    ncp_buffer.hpp                                  \
     ncp_config.h                                    \
     ncp_spi.cpp                                     \
     ncp_spi.hpp                                     \
     ncp_uart.cpp                                    \
     ncp_uart.hpp                                    \
-    spinel.c                                        \
-    spinel.h                                        \
-    spinel_decoder.cpp                              \
-    spinel_decoder.hpp                              \
-    spinel_encoder.cpp                              \
-    spinel_encoder.hpp                              \
-    spinel_platform.h                               \
     $(NULL)
 
 if OPENTHREAD_ENABLE_NCP_VENDOR_HOOK
@@ -138,26 +141,5 @@ libopenthread_rcp_a_SOURCES                       = \
 
 include_HEADERS                                   = \
     $(NULL)
-
-if OPENTHREAD_BUILD_TESTS
-
-check_PROGRAMS                                    = spinel-test
-spinel_test_SOURCES                               = spinel.c
-spinel_test_CFLAGS                                = \
-    $(COMMON_CPPFLAGS)                              \
-    -DSPINEL_SELF_TEST=1                            \
-    -D_GNU_SOURCE                                   \
-    -I$(top_srcdir)/src/core                        \
-    -I$(top_srcdir)/include                         \
-    $(NULL)
-
-TESTS                                             = spinel-test
-
-install-headers: install-includeHEADERS
-
-if OPENTHREAD_BUILD_COVERAGE
-CLEANFILES                                        = $(wildcard *.gcda *.gcno)
-endif # OPENTHREAD_BUILD_COVERAGE
-endif # OPENTHREAD_BUILD_TESTS
 
 include $(abs_top_nlbuild_autotools_dir)/automake/post.am
