+++ src/utils/spinel_encrypter.hpp
