+++ src/utils/spinel_platform.h
@@ -49,8 +49,6 @@
  *
  */
 
-#include "openthread-core-config.h"
-
 #include <stdarg.h>
 #include <stdbool.h>
 #include <stdint.h>
